# client/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "vibe-web-starter",
    "description": "\n        # AI 데이터 분석 웹 서비스 템플릿\n\n        FastAPI + SQLAlchemy 기반의 확장 가능한 백엔드 아키텍처\n\n        ## 주요 기능\n\n        - **도메인 플러그인 구조**: 새로운 도메인을 쉽게 추가 가능\n        - **계층화된 아키텍처**: Router → Service → Provider/Calculator/Formatter\n        - **타입 안전성**: Pydantic v2 + SQLAlchemy 2.0\n        - **비동기 처리**: async/await 기반\n\n        ## 아키텍처\n\n        ```\n        Router (FastAPI)\n            ↓\n        Service (비즈니스 로직 조율)\n            ↓\n        ├─ Provider (데이터 조회)\n        ├─ Calculator (계산/분석)\n        └─ Formatter (응답 포맷팅)\n        ```\n\n        ## 새 도메인 추가 방법\n\n        1. `server/app/examples/` 또는 `server/app/domain/` 에 새 디렉토리 생성\n        2. Provider, Calculator, Formatter, Service 구현\n        3. `server/app/api/v1/endpoints/` 에 라우터 추가\n        4. `server/app/api/v1/router.py` 에 라우터 등록\n        ",
    "version": "0.1.0"
  },
  "paths": {
    "/core/health": {
      "get": {
        "tags": [
          "core"
        ],
        "summary": "헬스체크",
        "description": "서비스의 상태를 확인합니다.\n\n    **사용 사례:**\n    - 로드밸런서의 헬스체크 타겟\n    - 모니터링 툴 (Datadog, New Relic 등)\n    - Kubernetes Liveness Probe\n\n    **응답 상태:**\n    - `ok`: 정상 작동\n    - `degraded`: 일부 기능 제한 (예: DB 연결 불안정)\n    - `error`: 서비스 이용 불가",
        "operationId": "health_check_core_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Check Core Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/core/version": {
      "get": {
        "tags": [
          "core"
        ],
        "summary": "버전 정보",
        "description": "애플리케이션의 버전 정보를 반환합니다.\n\n    **사용 사례:**\n    - 배포 후 버전 확인\n    - 프론트엔드에서 API 버전 체크\n    - 디버깅 시 환경 확인\n\n    **활용 예시:**\n    ```bash\n    # 배포 확인\n    curl https://api.example.com/core/version\n\n    # 프론트엔드에서 버전 불일치 감지\n    if (apiVersion !== expectedVersion) {\n      alert('새 버전이 배포되었습니다. 페이지를 새로고침해주세요.');\n    }\n    ```",
        "operationId": "version_info_core_version_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Version Info Core Version Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sample/": {
      "get": {
        "tags": [
          "sample"
        ],
        "summary": "샘플 데이터 목록 조회",
        "description": "전체 샘플 데이터 목록을 반환합니다.",
        "operationId": "list_sample_api_v1_sample__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_SampleDataResponse__"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "sample"
        ],
        "summary": "샘플 데이터 생성",
        "description": "새로운 샘플 데이터를 생성합니다.",
        "operationId": "create_sample_api_v1_sample__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SampleDataCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SampleDataResponse_"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sample/{item_id}": {
      "get": {
        "tags": [
          "sample"
        ],
        "summary": "샘플 데이터 단건 조회",
        "description": "ID로 샘플 데이터를 조회합니다.",
        "operationId": "get_sample_api_v1_sample__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SampleDataResponse_"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "sample"
        ],
        "summary": "샘플 데이터 수정",
        "description": "기존 샘플 데이터를 수정합니다.",
        "operationId": "update_sample_api_v1_sample__item_id__put",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SampleDataUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SampleDataResponse_"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sample"
        ],
        "summary": "샘플 데이터 삭제",
        "description": "샘플 데이터를 삭제합니다.",
        "operationId": "delete_sample_api_v1_sample__item_id__delete",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/docs": {
      "get": {
        "tags": [
          "documentation"
        ],
        "summary": "문서 파일 조회",
        "description": "프로젝트 문서 파일(Markdown)을 조회합니다.",
        "operationId": "get_document_api_v1_docs_get",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "문서 파일 경로 (예: /ARCHITECTURE.md)",
              "title": "Path"
            },
            "description": "문서 파일 경로 (예: /ARCHITECTURE.md)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/docs/list": {
      "get": {
        "tags": [
          "documentation"
        ],
        "summary": "문서 목록 조회",
        "description": "사용 가능한 문서 목록을 조회합니다.",
        "operationId": "list_documents_api_v1_docs_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "items": {},
                    "type": "array"
                  },
                  "type": "object",
                  "title": "Response List Documents Api V1 Docs List Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system/db-check": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "데이터베이스 연결 테스트",
        "description": "Supabase 데이터베이스 연결 상태를 확인하고 테스트 메시지를 반환합니다.",
        "operationId": "check_database_connection_api_v1_system_db_check_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DBCheckResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "API 헬스체크",
        "description": "API v1의 전체 상태를 확인합니다.",
        "operationId": "health_check_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Check Api V1 Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": [
          "root"
        ],
        "summary": "루트 엔드포인트",
        "description": "루트 엔드포인트\n\nAPI 기본 정보를 반환합니다.",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Root  Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResponse_SampleDataResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "요청 성공 여부"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SampleDataResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "응답 데이터"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "성공 메시지"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "에러 메시지"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ApiResponse[SampleDataResponse]"
      },
      "ApiResponse_list_SampleDataResponse__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "요청 성공 여부"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SampleDataResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "응답 데이터"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "성공 메시지"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "에러 메시지"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ApiResponse[list[SampleDataResponse]]"
      },
      "DBCheckResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "연결 성공 여부"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "연결 메시지"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "응답 시각"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "timestamp"
        ],
        "title": "DBCheckResponse",
        "description": "DB 연결 체크 응답"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SampleDataCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "데이터 이름 (1~255자)"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "상세 설명"
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "수치 값"
          },
          "score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Score",
            "description": "점수 (0.0 ~ 1.0)"
          }
        },
        "type": "object",
        "required": [
          "name",
          "value"
        ],
        "title": "SampleDataCreate",
        "description": "샘플 데이터 생성 요청 스키마",
        "examples": [
          {
            "description": "2026년 1분기 매출 실적",
            "name": "매출 데이터",
            "score": 0.85,
            "value": 1500000.0
          }
        ]
      },
      "SampleDataResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "고유 식별자"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "데이터 이름"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "상세 설명"
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "수치 값"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score",
            "description": "점수 (0.0 ~ 1.0)"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "생성 시각 (ISO 8601)"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "수정 시각 (ISO 8601)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "value",
          "created_at",
          "updated_at"
        ],
        "title": "SampleDataResponse",
        "description": "샘플 데이터 응답 스키마",
        "examples": [
          {
            "created_at": "2026-03-26T09:00:00+09:00",
            "description": "2026년 1분기 매출 실적",
            "id": 1,
            "name": "매출 데이터",
            "score": 0.85,
            "updated_at": "2026-03-26T09:00:00+09:00",
            "value": 1500000.0
          }
        ]
      },
      "SampleDataUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "데이터 이름 (1~255자)"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "상세 설명"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value",
            "description": "수치 값"
          },
          "score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Score",
            "description": "점수 (0.0 ~ 1.0)"
          }
        },
        "type": "object",
        "title": "SampleDataUpdate",
        "description": "샘플 데이터 수정 요청 스키마 (부분 수정 가능)",
        "examples": [
          {
            "name": "수정된 매출 데이터",
            "value": 1800000.0
          }
        ]
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}